requirements.txt
langchain
langgraph
langchain-groq
python-dotenv

pdfplumber
pdf2image
pytesseract
Pillow

rapidfuzz

numpy